Continuation user integration tests

+++ Tests/EM.IntegrationTests/Setup/DatabaseFixture.cs
@@ -1,34 +1,73 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Data.SqlClient;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoFixture;
 using Dapper;
 using EventManagement.Application.Models.Authorization;
+using EventManagement.Application.Models.Enums.Auth;
 using Microsoft.AspNetCore.Identity;
 
 namespace EM.IntegrationTests.Setup
 {
     public class DatabaseFixture
     {
+        private const string Connection =
+            "Server=host.docker.internal,1440;Database=EventManagementTests;User Id=sa;Password=Password_123BD;";
+
         private readonly IPasswordHasher<User> _passwordHasher;
         private readonly Fixture _fixture;
         private int _randomInt;
 
         public DatabaseFixture()
         {
+            this._fixture = new Fixture();
+            this._passwordHasher = new PasswordHasher<User>();
         }
 
-        public DatabaseFixture(IPasswordHasher<User> passwordHasher)
+        public async Task SetCustomUser(User user, string password)
         {
-            this._fixture = new Fixture();
-            this._passwordHasher = passwordHasher;
+            user.PasswordHash = this._passwordHasher.HashPassword(user, password);
+            var param = this.GetUserParameters(user);
+
+            await using var connection = new SqlConnection(Connection);
+            await connection.OpenAsync();
+            await using var transaction = connection.BeginTransaction(IsolationLevel.ReadUncommitted);
+
+            try
+            {
+                var success = await connection.ExecuteAsync("user_createNewUser_I", param,
+                    commandType: CommandType.StoredProcedure, transaction: transaction);
+
+                if (success <= 0)
+                {
+                    throw new ArgumentException("Threw an exception while inserting users.");
+                }
+                var identifier = param.Get<int?>("@new_identity");
+                var parameters = new DynamicParameters();
+
+                parameters.Add("@userId", identifier);
+                parameters.Add("@role", (int) Roles.User);
+
+                await connection.ExecuteAsync("user_addToRole_I", parameters, commandType: CommandType.StoredProcedure,
+                    transaction: transaction);
+
+                transaction.Commit();
+            }
+            catch
+            {
+                transaction.Rollback();
+                throw;
+            }
+
         }
 
         public async Task UserSeedData()
         {
             this._randomInt = this.GetRandomInt();
+            var userList = new List<User>();
             for (var i = 0; i < this._randomInt; i++)
             {
                 var userTest = new User()
@@ -43,13 +82,44 @@ namespace EM.IntegrationTests.Setup
                 };
                 var passwordTest = userTest.UserName + "$" + "123";
                 userTest.PasswordHash = this._passwordHasher.HashPassword(userTest, passwordTest);
+                userList.Add(userTest);
+            }
+
+            await using var connection = new SqlConnection(Connection);
+            await connection.OpenAsync();
+            await using var transaction = connection.BeginTransaction(IsolationLevel.ReadUncommitted);
+            try
+            {
+                foreach (var user in userList)
+                {
+                    var param = this.GetUserParameters(user);
+                    var identifier = await connection.ExecuteAsync("user_createNewUser_I", param,
+                        commandType: CommandType.StoredProcedure, transaction: transaction);
+
+                    if (identifier <= 0)
+                    {
+                        throw new ArgumentException("Threw an exception while inserting users.");
+                    }
+
+                    var parameters = new DynamicParameters();
+
+                    parameters.Add("@userId", identifier);
+                    parameters.Add("@role", (int) Roles.User);
+
+                    await connection.ExecuteAsync("user_addToRole_I", parameters, commandType: CommandType.StoredProcedure,
+                        transaction: transaction);
+                }
+
+                transaction.Commit();
+            }
+            catch
+            {
+                transaction.Rollback();
+                throw;
             }
         }
 
-        private int GetRandomInt(int a = 1, int b = 10) => new Random().Next(a, b);
-
-
-        public async Task DeleteData(string tableName = null)
+        public async Task DeleteData(string tableName = null, string constScript = null)
         {
             var tablesList = new List<string>
             {
@@ -58,22 +128,24 @@ namespace EM.IntegrationTests.Setup
                 "Performers", "Events", "ApplicationUsers"
             };
 
-            var script = string.Empty;
-            if (string.IsNullOrEmpty(tableName))
+            string script;
+            if (string.IsNullOrEmpty(tableName) && constScript == null)
             {
                 script = @"BEGIN TRANSACTION 
-                       DELETE FROM Opinions
-                       DELETE FROM EventImages
-                       DELETE FROM UserRoles
-                       DELETE FROM RefreshTokens
-                       DELETE FROM PerformanceProposals
-                       DELETE FROM EventApplications
-                       DELETE FROM Performers
-                       DELETE FROM [Events]
-                       DELETE FROM ApplicationUsers
-                       COMMIT TRANSACTION";
+                           DELETE FROM Opinions
+	                       DELETE FROM EventImages
+	                       DELETE FROM UserRoles 
+	                       WHERE UserId != 1 
+	                       DELETE FROM RefreshTokens
+	                       DELETE FROM PerformanceProposals
+	                       DELETE FROM EventApplications
+	                       DELETE FROM Performers
+	                       DELETE FROM [Events]
+	                       DELETE FROM ApplicationUsers 
+	                       WHERE Email != 'SuperUser@test.com'  
+                           COMMIT TRANSACTION";
             }
-            else
+            else if(!string.IsNullOrEmpty(tableName) && constScript == null)
             {
                 var result =
                     tablesList.FirstOrDefault(_ => _.Equals(tableName, StringComparison.CurrentCultureIgnoreCase));
@@ -84,17 +156,38 @@ namespace EM.IntegrationTests.Setup
 
                 script = $"DELETE FROM {result}";
             }
+            else
+            {
+                script = constScript;
+            }
 
             if (string.IsNullOrEmpty(script))
             {
                 throw new ArgumentNullException(nameof(script));
             }
 
-            await using var connection = new SqlConnection(
-                "Server=host.docker.internal,1440;Database=EventManagementTests;User Id=sa;Password=Password_123BD;");
+            await using var connection = new SqlConnection(Connection);
             await connection.OpenAsync();
 
             await connection.ExecuteAsync(script);
+        }
+ 
+
+        private int GetRandomInt(int a = 1, int b = 10) => new Random().Next(a, b);
+
+        private DynamicParameters GetUserParameters(User user)
+        {
+            var param = new DynamicParameters();
+
+            param.Add("@firstName", user.FirstName);
+            param.Add("@lastName", user.LastName);
+            param.Add("@userName", user.UserName);
+            param.Add("@email", user.Email);
+            param.Add("@phoneNumber", user.PhoneNumber);
+            param.Add("@passwordHash", user.PasswordHash);
+            param.Add("@new_identity", -1, DbType.Int32, ParameterDirection.Output);
+
+            return param;
         }
     }
 }
